pyproject.toml
[project]
name = "shooting-session-dataset-pipeline"
version = "0.1.0"
description = "Dataset creation pipeline for ball and goal+cone YOLO models"
requires-python = ">=3.10"
dependencies = [
    "ultralytics>=8.3.0",
    "opencv-python>=4.9.0.80",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0",
    "rich>=13.0.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["dataset_pipeline", "training"]
